+++ src/compas_fab/backends/ros/__init__.py
@@ -2,6 +2,7 @@ from .client import RosClient
 from .exceptions import RosError
 from .exceptions import RosValidationError
 from .fileserver_loader import RosFileServerLoader
+from .http_fileserver_loader import HttpFileServerLoader
 from .messages import ROSmsg
 from .planner import MoveItPlanner
 
@@ -9,6 +10,7 @@ __all__ = [
     "ROSmsg",
     "RosClient",
     "RosFileServerLoader",
+    "HttpFileServerLoader",
     "RosError",
     "RosValidationError",
     "MoveItPlanner",